part2/hbnb/requirements.txt
flask
flask-restx
flask-bcrypt  